requirements.txt
beautifulsoup4==4.14.3
feedparser==6.0.12
pandas==3.0.2
protobuf==7.34.1
python-dotenv==1.2.2
